feat: move entity type template to main entityId route (#496)
* feat: migrate profile template from /template to the main entityId route

* fix(hack): stub change to trigger CI on new job requirements

* fix: run github workflows on stream/social branch
