+++ .github/workflows/e2e.yml
@@ -6,9 +6,11 @@ on:
   push:
     branches:
       - master
+      - stream/social
   pull_request:
     branches:
       - master
+      - stream/social
 
 jobs:
   playwright:
