+++ .github/workflows/test.yml
@@ -4,9 +4,11 @@ on:
   push:
     branches:
       - master
+      - stream/social
   pull_request:
     branches:
       - master
+      - stream/social
 
 jobs:
   test:
