+++ apps/web/app/space/[id]/[entityId]/activity/layout.tsx
