+++ apps/web/app/space/[id]/[entityId]/education/page.tsx
