+++ apps/web/app/space/[id]/[entityId]/profile-entity-server-container.tsx
@@ -6,7 +6,7 @@ import { ServerSideEnvParams } from '~/core/types';
 import { Entity } from '~/core/utils/entity';
 
 // import { setOnboardingDismissedCookie } from '~/partials/profile/actions';
-import { ProfilePageComponent } from './profile-client-page';
+import { ProfilePageComponent } from './profile-entity-page';
 
 export const runtime = 'edge';
 
@@ -15,13 +15,9 @@ interface Props {
   searchParams: ServerSideEnvParams;
 }
 
-export async function ProfileServerPage({ params, searchParams }: Props) {
+export async function ProfileEntityServerContainer({ params, searchParams }: Props) {
   const env = cookies().get(Params.ENV_PARAM_NAME)?.value;
   const config = Params.getConfigFromParams(searchParams, env);
-
-  // @TODO: Disabling cookie interactions for now until we get later on in the social
-  // work. This is so we can test onboarding feedback more frequently.
-  // const hasDismissedOnboarding = cookies().get(Cookie.HAS_DISMISSED_PERSONAL_SPACE_ONBOARDING_KEY)?.value === 'true';
 
   const person = await Subgraph.fetchEntity({ id: params.entityId, endpoint: config.subgraph });
 
@@ -44,5 +40,5 @@ export async function ProfileServerPage({ params, searchParams }: Props) {
     coverUrl: Entity.cover(person.triples),
   };
 
-  return <ProfilePageComponent {...profile} spaceId={params.id} hasDismissedOnboarding={false} />;
+  return <ProfilePageComponent id={profile.id} triples={profile.triples} spaceId={params.id} />;
 }
