+++ apps/web/app/space/[id]/[entityId]/triples/page.tsx
