+++ apps/web/app/space/[id]/[entityId]/work/page.tsx
