+++ apps/web/partials/profile/activity-space-filter.tsx
@@ -3,7 +3,7 @@
 import { SYSTEM_IDS } from '@geogenesis/ids';
 import Image from 'next/legacy/image';
 import Link from 'next/link';
-import { useRouter, useSearchParams } from 'next/navigation';
+import { useSearchParams } from 'next/navigation';
 
 import * as React from 'react';
 
@@ -27,7 +27,6 @@ export function ActivitySpaceFilter({ entityId, spaceId }: Props) {
   const initialSpace = spaces.find(space => space.id === selectedSpaceId);
   const initialName = initialSpace?.attributes[SYSTEM_IDS.NAME];
 
-  const router = useRouter();
   const [open, onOpenChange] = React.useState(false);
   const [name, setName] = React.useState('All');
 
@@ -64,17 +63,17 @@ export function ActivitySpaceFilter({ entityId, spaceId }: Props) {
           {name}
         </SmallButton>
       }
-      className="flex flex-col max-h-[300px] max-w-[250px] overflow-y-auto"
+      className="flex max-h-[300px] max-w-[250px] flex-col overflow-y-auto"
     >
       {spacesWithAll.map(space => (
         <Link
           href={NavUtils.toProfileActivity(spaceId, entityId, space.id === 'all' ? undefined : space.id)}
           onClick={() => onSelect(space.id)}
           key={space.id}
-          className="text-button p-3 bg-white text-grey-04 hover:text-text hover:bg-bg transition-colors duration-75 flex gap-2 w-full"
+          className="flex w-full gap-2 bg-white p-3 text-button text-grey-04 transition-colors duration-75 hover:bg-bg hover:text-text"
         >
           {space.attributes[SYSTEM_IDS.IMAGE_ATTRIBUTE] && (
-            <div className="relative w-3 h-3 rounded-xs overflow-hidden mt-[4.5px]">
+            <div className="relative mt-[4.5px] h-3 w-3 overflow-hidden rounded-xs">
               <Image src={getImagePath(space.attributes[SYSTEM_IDS.IMAGE_ATTRIBUTE])} layout="fill" objectFit="cover" />
             </div>
           )}
